created shared SpaceRootLayout component
