+++ packages/gitbook/src/components/SpaceRootLayout/globals.css
